refactor: java doc & handler interface
