+++ app/src/main/java/com/codesoom/assignment/App.java
@@ -1,11 +1,12 @@
 package com.codesoom.assignment;
 
-import com.sun.net.httpserver.HttpHandler;
+import com.codesoom.assignment.impl.AssignmentHttpHandler;
 import com.sun.net.httpserver.HttpServer;
 import java.io.IOException;
 import java.net.InetSocketAddress;
 
 public class App {
+    private static final IHttphandler httphandler = new AssignmentHttpHandler();
 
     public String getGreeting() {
         return "Hello World!";
@@ -14,8 +15,7 @@ public class App {
     public static void main(String[] args) throws IOException {
         InetSocketAddress address = new InetSocketAddress(8000);
         HttpServer httpServer = HttpServer.create(address, 0);
-        HttpHandler handler = new AssignmentHttpHandler();
-        httpServer.createContext("/", handler);
+        httpServer.createContext("/", httphandler);
         httpServer.start();
     }
 }
